Structure mobile navigation around native stack
