+++ apps/mobile/src/screens/new/add-project/repository.tsx
@@ -1,10 +1,10 @@
-import { Stack, useLocalSearchParams } from "../../../navigation/router";
+import { NativeStackScreenOptions, useRouteParams } from "../../../navigation/native-stack-header";
 import { addProjectRemoteSourceLabel } from "@t3tools/client-runtime/operations/projects";
 
 import { AddProjectRepositoryScreen } from "../../../features/projects/AddProjectScreen";
 
 export default function AddProjectRepositoryRoute() {
-  const params = useLocalSearchParams<{ source?: string | string[] }>();
+  const params = useRouteParams<{ source?: string | string[] }>();
   const source = Array.isArray(params.source) ? params.source[0] : params.source;
   const title =
     source === "github" ||
@@ -16,7 +16,7 @@ export default function AddProjectRepositoryRoute() {
 
   return (
     <>
-      <Stack.Screen options={{ title }} />
+      <NativeStackScreenOptions options={{ title }} />
       <AddProjectRepositoryScreen />
     </>
   );
