--- apps/mobile/src/screens/settings/archive.tsx
@@ -1,3 +1,0 @@
-import { ArchivedThreadsRouteScreen } from "../../features/archive/ArchivedThreadsRouteScreen";
-
-export default ArchivedThreadsRouteScreen;
