--- apps/mobile/src/screens/settings/environment-new.tsx
@@ -1,1 +1,0 @@
-export { default } from "../connections/new";
