--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/files/[...path].tsx
@@ -1,5 +1,0 @@
-import { ThreadFileScreen } from "../../../../../features/files/ThreadFilesRouteScreen";
-
-export default function ThreadFileRoute() {
-  return <ThreadFileScreen />;
-}
