--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/files/index.tsx
@@ -1,5 +1,0 @@
-import { ThreadFilesTreeScreen } from "../../../../../features/files/ThreadFilesRouteScreen";
-
-export default function ThreadFilesIndexRoute() {
-  return <ThreadFilesTreeScreen />;
-}
