--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/git-confirm.tsx
@@ -1,5 +1,0 @@
-import { GitConfirmSheet } from "../../../../features/threads/git/GitConfirmSheet";
-
-export default function GitConfirmRoute() {
-  return <GitConfirmSheet />;
-}
