--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/git/branches.tsx
@@ -1,5 +1,0 @@
-import { GitBranchesSheet } from "../../../../../features/threads/git/GitBranchesSheet";
-
-export default function GitBranchesRoute() {
-  return <GitBranchesSheet />;
-}
