--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/git/commit.tsx
@@ -1,5 +1,0 @@
-import { GitCommitSheet } from "../../../../../features/threads/git/GitCommitSheet";
-
-export default function GitCommitRoute() {
-  return <GitCommitSheet />;
-}
