--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/git/index.tsx
@@ -1,5 +1,0 @@
-import { GitOverviewSheet } from "../../../../../features/threads/git/GitOverviewSheet";
-
-export default function GitRoute() {
-  return <GitOverviewSheet />;
-}
