--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/index.tsx
@@ -1,5 +1,0 @@
-import { ThreadRouteScreen } from "../../../../features/threads/ThreadRouteScreen";
-
-export default function ThreadRoute() {
-  return <ThreadRouteScreen />;
-}
