--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/review-comment.tsx
@@ -1,5 +1,0 @@
-import { ReviewCommentComposerSheet } from "../../../../features/review/ReviewCommentComposerSheet";
-
-export default function ReviewCommentRoute() {
-  return <ReviewCommentComposerSheet />;
-}
