--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/review.tsx
@@ -1,10 +1,0 @@
-import { ReviewHighlighterProvider } from "../../../../features/review/ReviewHighlighterProvider";
-import { ReviewSheet } from "../../../../features/review/ReviewSheet";
-
-export default function ReviewRoute() {
-  return (
-    <ReviewHighlighterProvider>
-      <ReviewSheet />
-    </ReviewHighlighterProvider>
-  );
-}
