--- apps/mobile/src/screens/threads/[environmentId]/[threadId]/terminal.tsx
@@ -1,5 +1,0 @@
-import { ThreadTerminalRouteScreen } from "../../../../features/terminal/ThreadTerminalRouteScreen";
-
-export default function ThreadTerminalRoute() {
-  return <ThreadTerminalRouteScreen />;
-}
